Fix: 배송비 유효성 검사 체크

+++ src/main/java/shop/s5g/shop/dto/delivery/DeliveryResponseDto.java
@@ -8,7 +8,7 @@ public record DeliveryResponseDto(
     String address,
     LocalDate receivedDate,
     LocalDate shippingDate,
-    int fee,
+    long fee,
     String invoiceNumber,
     String receiverName,
     String status
